% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fide_to_country_name.R
\name{fide_to_country_name}
\alias{fide_to_country_name}
\title{Fide to country name}
\usage{
fide_to_country_name(fc)
}
\arguments{
\item{fc}{Fide country code}
}
\description{
Convert Fide country codes to country name
}
\examples{
fide_to_country_name(c("ENG","USA","RUS","FID","GER"))
}
\keyword{Fide}
\keyword{country}
\keyword{name}

