% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fide_to_ioc.R
\name{fide_to_ioc}
\alias{fide_to_ioc}
\title{Fide to ioc}
\usage{
fide_to_ioc(fc)
}
\arguments{
\item{fc}{Fide country code}
}
\description{
Convert Fide country codes to ioc
}
\examples{
fide_to_ioc(c("ENG","USA","RUS"))
}
\keyword{Fide}
\keyword{code}
\keyword{country}
\keyword{ioc}

